kde: make active task manager shortcuts work
Activate task manager shortcuts work only when the task manager is in a
plasma panel (not latte-dock). I've created a empty plasma panel with
only the task manager widget. It is hidden behind the top latte panel.

Bonus: latte panel and desktop files won't overlap, thanks to the
presence of the plasma panel.