kde: use plasma panel as top panel
The workaround attempted in a35838bc5117eba258afeb36ca9748faf67a01be
didn't work so well. Apparently the icon-only task manager on the latte
dock and the icon-only task manager in the hidden plasma panel had
different order, so the shortcuts were misleading.

Decision: just use a plasma panel... the only disadvantage is that I
can't have the digital clock centered, but that's not a big deal.