vim: replace mhinz/vim-signify with lewis6991/gitsigns.nvim
I was having sometimes (typically after having the same neovim instance
open for a while) the error "too many highlight groups". From
https://www.reddit.com/r/neovim/comments/qn8s0u/e849_too_many_highlight_syntax_groups/
it seems a bug in Neovim, but I don't want to upgrade to neovim nightly 7.0.
People also say that it is vim-signify that triggers the problem. Well,
it's high time that we jump to a more modern git signs plugin written in
lua, i.e. lewis6991/gitsigns.nvim.