vifm: remove invalid config
Seems that this version of vifm (which I got presumably since the last
NixOS update that I did)

```
Version: 0.13
Git info: built out of repository
Compiled at: Apr  3 2023 22:19:55
```

doesn't accept this config anymore. In particular, `source()` now
complains if the file being sourced is does not exist.
Anyway not a problem for me.