nixos: fix kde config for custom shortcuts
This fix was long due since the upgrade to Plasma 6. But actually also
the new config doesn't work... If I remove the shortcuts from the GUI
and re-run home-manager to write the correct things in the KDE configuration
file, the shortcut still don't show up. They must be created via GUI.