markdown: switch from markdown-cli2 to markdown-cli
The later supports finding configuration in ancestor directories.
Useful for having a default configuration.
NOTE that configuration in ancestor directories must be in JSON or RC
format. JSONC is not supported. But I want to keep comments in my
markdownlint config. That's why I implemented a simple Nix function to strip
JSONC comments.